Smaller version: 4 lines removed
When reading the article I directly see that the `for` in `Match` could be optimized.
Then I see that the same things could be done with `matchStar`.

I think the first change (in `Match`) is OK and its still readable.
For the second one (in `matchStar`), it might remove some readability ?